Moved the logic of LoginController over to LoginLogic
