+++ src/main/java/UnemployedVoodooFamily/GUI/LoginController.java
@@ -1,8 +1,6 @@
 package UnemployedVoodooFamily.GUI;
 
-import ch.simas.jtoggl.JToggl;
-import ch.simas.jtoggl.User;
-import javafx.application.Platform;
+import UnemployedVoodooFamily.Logic.LoginLogic;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.PasswordField;
@@ -10,20 +8,18 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.KeyEvent;
 
-import java.io.IOException;
-
 public class LoginController {
 
     @FXML
     private Button submitBtn;
     @FXML
+    private ImageView bufferImg;
+    @FXML
     private TextField emailField;
     @FXML
     private PasswordField passwordField;
-    @FXML
-    private ImageView bufferImg;
 
-    private JToggl jToggl;
+    private LoginLogic loginLogic = new LoginLogic();
 
 
     public void initialize() {
@@ -31,62 +27,18 @@ public class LoginController {
     }
 
     private void setKeyAndClickListeners() {
-        submitBtn.setOnAction(event -> login());
+        submitBtn.setOnAction(event -> loginWithCredentials());
     }
 
-
-    private void login() {
+    public void loginWithCredentials() {
         bufferImg.setVisible(true);
-        attemptAuthentication(emailField.getText(), passwordField.getText());
+        loginLogic.attemptAuthentication(emailField.getText(), passwordField.getText());
+        bufferImg.setVisible(false);
     }
 
-
-    private void attemptAuthentication(String username, String password){
-        // Run this thread to avoid UnemployedVoodooFamily.GUI freezing
-        Thread toggleThread = new Thread(() -> {
-            jToggl = new JToggl(username, password);
-            jToggl.switchLoggingOn();
-            if(isLoggedIn()){
-                Platform.runLater(() -> {
-                    try{
-                        new GUIBaseController().start();
-                    }
-                    catch(IOException e) {
-                        e.printStackTrace();
-                    }
-                });
-            }
-            bufferImg.setVisible(false);
-
-            });
-        toggleThread.start();
-    }
-
-
-    private boolean isLoggedIn() {
-        String userString;
-        boolean loggedIn = false;
-        try{
-            User user = jToggl.getCurrentUser();
-            userString = user.toString();
-            if(userString.contains("api_token")) {
-                loggedIn = true;
-            }
-            else {
-                loggedIn = false;
-            }
-        }
-        catch(RuntimeException t){
-            System.out.println("\nWrong username or password");
-        }
-        return loggedIn;
-    }
-
-
-    @FXML
     public void buttonPressedListener(KeyEvent e) {
         if(e.getCode().toString().equals("ENTER")) {
-            login();
+            loginWithCredentials();
         }
     }
 }
